refactor(chat): localize display model
