+++ docs/architecture/chat-display-model-boundary/tasks.md
@@ -1,0 +1,7 @@
+# 任务清单
+
+- [x] 审计展示模型、renderability policy 与全部 source/test 消费者。
+- [x] 在 `features/chat-page/model/` 建立纯展示模型模块。
+- [x] 原子更新 renderer source、tests 与 fixtures 的 import。
+- [x] 删除旧 component-local 模块并确认无残留引用。
+- [x] 执行 format、i18n、lint、typecheck 与 targeted renderer tests。
