+++ src/renderer/src/components/message/MessageBlockAudio.vue
@@ -35,7 +35,7 @@ import { computed, ref } from 'vue'
 import { Icon } from '@iconify/vue'
 import { useI18n } from 'vue-i18n'
 import { Spinner } from '@shadcn/components/ui/spinner'
-import type { DisplayAssistantMessageBlock } from '@/components/chat/messageListItems'
+import type { DisplayAssistantMessageBlock } from '@/features/chat-page/model/displayMessage'
 
 const keyMap = {
   'mcp.sampling.contentType.audio': 'Audio',
