+++ src/renderer/src/components/message/MessageBlockError.vue
@@ -47,7 +47,7 @@
 import { useI18n } from 'vue-i18n'
 import { Icon } from '@iconify/vue'
 import { computed, ref, useId } from 'vue'
-import type { DisplayAssistantMessageBlock } from '@/components/chat/messageListItems'
+import type { DisplayAssistantMessageBlock } from '@/features/chat-page/model/displayMessage'
 const { t } = useI18n()
 
 const props = defineProps<{
