+++ src/renderer/src/components/message/MessageBlockQuestionRequest.vue
@@ -28,7 +28,7 @@
 <script setup lang="ts">
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
-import type { DisplayAssistantMessageBlock } from '@/components/chat/messageListItems'
+import type { DisplayAssistantMessageBlock } from '@/features/chat-page/model/displayMessage'
 import type { QuestionOption } from '@shared/types/core/question'
 
 const props = defineProps<{
