+++ src/renderer/src/components/message/messageActivityGroups.ts
@@ -1,4 +1,4 @@
-import type { DisplayAssistantMessageBlock } from '@/components/chat/messageListItems'
+import type { DisplayAssistantMessageBlock } from '@/features/chat-page/model/displayMessage'
 
 export type AssistantRenderItem =
   | {
